Move functional utilities out of `H5WasmApi` class
